5th test successfully need to check Bigdecimal issue

+++ src/main/java/com/example/demo/repositories/CustomMaintenancePlanRepository.java
@@ -3,6 +3,7 @@ package com.example.demo.repositories;
 import com.example.demo.utils.Pair;
 import org.springframework.data.jpa.repository.Query;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface CustomMaintenancePlanRepository {
@@ -10,6 +11,6 @@ public interface CustomMaintenancePlanRepository {
     //@Query("select new com.example.demo.utils.Pair<Integer,Long>() from MaintenancePlan mp,mp.yearOfAction f,mp.tasks.")
     //List<Pair<Integer, Long>> findCorrectiveRepairsByYearForPeriod(int year, int thisyear);
 
-    List<Pair<Integer, Long>> findCorrectiveRepairCostsByYear(int year, int thisyear);
+    List<Pair<Integer, BigDecimal>> findCorrectiveRepairCostsByYear(int year, int thisyear);
 
 }
